Add script to parse old news entries and convert to markdown
The script should be executed from the `data` directory. It will read
the contents of the `data/data.json` file, which contains the news
entries in JSON form as dumped from the original MySQL database. The
converted entries are written as markdown files to the `docs/news/posts`
directory. It also writes the file `data/urls.json` which contains a
mapping of the old to the new URLs.